Creates a first fucking view for both Android and iOS
